+++ CHANGES.rst
@@ -10,6 +10,7 @@ Features:
 
 Improvements:
  - Add option to choose default media source (#2763)
+ - Add option to choose default photo compression (#2763)
 
 Other changes:
  -